requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.0
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
whitenoise==6.11.0
